Overhaul build, add figwheel and pals
1. Bump deps into the year of our Lord 2018.

2. Use no optimizations in dev instead of whitespace.  This improves the
dev feedback loop but causes problems where the entrypoint of the app is
located at a dynamic path, courtesy of boot-cljs.  To that end:

3. Generate manifest.json and popup.html dynamically as a part of the build.
This allows us to produce different versions based on the whims of the
ClojureScript compiler and on our build profile.

I consider this to be a strict improvement over forking static assets between
dev/build and bypassing the intended entrypoint in dev like the upstream lein
variant of this project does.

4. Provide a figwheel-driven dev environment, with other tools to boot (haHAA).